Add OCR workpiles, removed ocr-factory

+++ src/policy/ocr-policy.c
@@ -33,15 +33,15 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <assert.h>
 
 #include "ocr-policy.h"
-#include "ocr-factory.h"
+#include "ocr-task-event.h"
 
 extern ocr_policy_domain_t * hc_policy_domain_constructor();
 
 ocr_policy_domain_t * newPolicy(ocr_policy_kind policyType,
-        int nb_workpiles,
-        int nb_workers,
-        int nb_executors,
-        int nb_schedulers) {
+        size_t nb_workpiles,
+        size_t nb_workers,
+        size_t nb_executors,
+        size_t nb_schedulers) {
     switch(policyType) {
     case OCR_POLICY_HC:
         return hc_policy_domain_constructor(nb_workpiles, nb_workers,
@@ -145,10 +145,10 @@ void hc_policy_domain_destroy(ocr_policy_domain_t * policy) {
     free(policy);
 }
 
-ocr_policy_domain_t * hc_policy_domain_constructor(int nb_workpiles,
-        int nb_workers,
-        int nb_executors,
-        int nb_schedulers) {
+ocr_policy_domain_t * hc_policy_domain_constructor(size_t nb_workpiles,
+        size_t nb_workers,
+        size_t nb_executors,
+        size_t nb_schedulers) {
     ocr_policy_domain_t * policy = (ocr_policy_domain_t *) malloc(sizeof(ocr_policy_domain_t));
     policy->nb_executors = nb_executors;
     policy->nb_workpiles = nb_workpiles;
